Integer I call you!

+++ kmserver-core/src/dao/sqls/kara.ts
@@ -120,7 +120,7 @@ DELETE FROM kara WHERE pk_kid = ANY ($1);
 export const getYears = (collectionClauses: string[]) => `
 SELECT DISTINCT
 	k.year,
-	COUNT(k2.pk_kid) AS karacount
+	COUNT(k2.pk_kid)::integer AS karacount
 FROM kara AS k
 LEFT JOIN kara k2 ON k2.pk_kid = k.pk_kid
 LEFT JOIN all_karas ak ON k2.pk_kid = ak.pk_kid
